requirements.txt
psycopg2-binary>=2.9.9
requests>=2.31.0
beautifulsoup4>=4.12.2
python-telegram-bot>=20.7
python-dotenv>=1.0.0
schedule>=1.2.0
lxml>=4.9.3
playwright>=1.40.0
pytesseract>=0.3.10
Pillow>=10.0.0
asyncio 